+++ security/sample/src/main/java/ru/surfstudio/android/security/sample/app/dagger/CustomAppComponent.kt
@@ -6,6 +6,7 @@ import ru.surfstudio.android.dagger.scope.PerApplication
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultAppComponent
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultAppModule
 import ru.surfstudio.android.sample.dagger.app.dagger.DefaultSharedPrefModule
+import ru.surfstudio.android.sample.dagger.app.dagger.NavigationModule
 import ru.surfstudio.android.security.sample.interactor.biometrics.BiometricsModule
 import ru.surfstudio.android.security.sample.interactor.cache.CacheModule
 import ru.surfstudio.android.security.sample.interactor.profile.ProfileInteractor
@@ -15,14 +16,17 @@ import ru.surfstudio.android.security.sample.interactor.storage.ApiKeyStorageWra
 import ru.surfstudio.android.security.session.SessionActivityCallback
 
 @PerApplication
-@Component(modules = [
-    DefaultAppModule::class,
-    DefaultSharedPrefModule::class,
-    CacheModule::class,
-    ApiKeyModule::class,
-    SessionModule::class,
-    BiometricsModule::class
-])
+@Component(
+    modules = [
+        DefaultAppModule::class,
+        DefaultSharedPrefModule::class,
+        NavigationModule::class,
+        CacheModule::class,
+        ApiKeyModule::class,
+        SessionModule::class,
+        BiometricsModule::class
+    ]
+)
 interface CustomAppComponent : DefaultAppComponent {
     fun apiKeyStorageWrapper(): ApiKeyStorageWrapper
     fun profileInteractor(): ProfileInteractor